Get Curve model/view roughly going
The codebase is in a chaotic state right now, but it does compile and
run stably. I'm comitting my work before I clean it up.

Curves can now be added to the database and retrieved from the database.
They can also be drawn to the screen with a neat visualizer (:D!).

Aside from lots of cleanup I would also like to bring in the Autotools.
I think the time has come. My pitiful excuse for a Makefile is just not
cutting the mustard anymore. :P