Get/save/update Curves
I had initially wanted this to be three commits (the dependencies go
update -> save -> get, as it happens). However, Git got the update/get
patch for curve.cpp very mixed-up because of my going too long between
commits while doing this, so I'm mixing them together.

This basically finishes CRUD for Curves. It provides an ActiveRecord-
like interface. You can get a Curve, run some stuff on it
in memory, then commit it to the database if you want with save(). You
can also retrieve a Curve from the DB with get(), change it around, then
save the changes with save() again.

My upcoming plan is to introduce a small DSL for curves and save
statements of that in the DB instead of the raw curve data.