Redesign Note range
Now it goes from Note(1, 0, 0) = 20Hz to Note(11, 0, 0) ~= 20,000 Hz.
Having used it a bit I really like it. Note(4, 31, 0)—Note(5, 31, 0) is
around the middle of the human hearing range this way which I think is
lovely. It's very intuitive and even seems like it might improve a
person's sense of absolute pitch somewhat to work with for a while.