Change Soidser default fundamental to Note(4, 00)
I realize there's a mistake here in that I didn't add
Soidser::inner_process() to the earlier "const" commit, so probably the
code won't compile between there and here. My apologies to anyone who
might be inconvenienced by this. I feel like the chances of it mattering
are small enough that I'm not inclined to go to the trouble of rebasing
everything, especially since I think it will be an easy mistake to spot.
Sorry if it really does cause anyone trouble, though (including if that
person is me in the future). This does make me more inclined, however,
to stash my unfinished changes and try compiling the code each time
before the next commit.

Anyway, this makes it so that the default Soidser fundamental is a
"decent default Note to test the sound of a curve with," in this case
Note(4, 00). Using a `fund` of `1.0` when constructing a `Soidser` has
the potential to be annoyingly slow (valgrind indicates it's due to the
many calls to `std::sin()`, unsurprisingly). Using Note(4, 00) instead
provides a dramatic improvement (for the common case when `part_spacing`
is `1`, 24,000 calls to `std::sin()` vs 75 calls to `std::sin()`). The
case where `fund` is `1.0` is mainly useful for the purpose of
visualization, which might suggest that the visual representation of a
Curve should be cached somewhere by CurveAlg and x/y term character
size, maybe even on disk in the database (the global user database?).