Actually play sound via SDL
Wavetable synthesis! The program shows the wave on the screen and plays
a chord using it through the speakers. This is actually fun to play
around with. :D

I had to really tear up the Audio/SDL area to get this working. There's
a real audio callback fn (or rather family of) and the callback actually
gets called now. Various bugs have been fixed. Best of all, it really
plays sound. Aside from the Curve DSL, which I might actually put off
for a bit until I really need to store curves in the DB, I think a good
next move would be to start designing a packet format for the "audio
server". If I approach it that way it will make the synthesizer part of
the program embeddable, and it will facilitate writing audio "routines"
with this program and then saving them as packet streams you can summon
up on command in a game or movie or something. Maybe I could even have
embeddable "stream emitters" that store a sort of stream template and
expose parameters you can feed in from a game engine or the like—that
would be a great step towards my dream of "audio materials". ^^