Have Curve() take an args struct, add sample_rate
I'm having Curve hold its own sample_rate for now, partially because it
will ease things and partially because I want to allow it to differ from
the output sample rate. In the course of setting that up, I felt like
keeping on with the overloaded constructors was a pain, so I think it's
time to use this approach instead.