It's doing something else really pretty
This is kind of an awkward place to commit but I just accidentally
stumbled on another really lovely behavior for the program. I'm not sure
what I did wrong yet, I just compiled, ran, and now am immediately
committing because I don't want to lose this regardless of whatever
else. It was perfectly producing the earlier rendition of the piece
until just a bit ago—I think like, just before I made the changes to
`PulseTrain`.