Use discrete time (may not be permanent)
Things are not exactly going according to plan. :/ For whatever reason,
I'm still getting behavior that is slightly different from the original
piece, but different enough to be obvious and vexing (sadly I don't like
this behavior as much). It starts out almost identically, but there
seems to be some sort of slight timing difference which results in the
two versions diverging after the first switch to the high chord or so,
varying more and more after that as time goes on.

At first, I suspected the change to discrete time, but actually what
evidence I have managed to gather does not suggest that the problem lies
there. Changing the coarseness of the discrete representation can result
in e.g. significant changes in pitch and overall speed if the
representation is made very coarse, because of discrepancies between the
real sample rate and the effective sample rate within the program as a
result of the coarse representation. However, the program still follows
the same new vaguely-wrong behavior as it would otherwise, just at a
different speed and pitch. In practice, moving to a discrete time
representation doesn't really seem to have hurt anything, although I
can't say that it's helped anything either at this point.

I think it's much more likely that the problem lies in the changes to
how the modules/flags are updated—specifically like, when the updates
vs. gets occur and how. I've been separating the modules' `get()` and
`update()` steps, whith `get()` being const, so that in the future I can
jump to any point in a piece and just take the `update()` and
`update()`-dependent `get()` steps to get there without having to spend
CPU time on the purely-auditory `get()` steps. The earlier divergence
from the original behavior, the one that I did like, occurred as I was
modifying `PulseTrain` to this effect. Interestingly, this behavior is
much closer to the original than that behavior was, but in any case I
may have introduced multiple bugs in the course of doing all this.

I think I'm going to walk the codebase back to a state where we can get
the original behavior, and then move forward more slowly and carefully.
This has been very frustrating! >:(